Simplify watch command
Let's drop some of those pipes in favour of a more native approach.
Also note the difference in namespace depending how the operator was
installed. We also need to filter for 'patterns' as sometimes other
operators could be in that namespace and we do not want to always match
the first line.