Add proper ARGOCD_CLUSTER_CONFIG_NAMESPACES to config environment in the ArgoCD subscription
Currently when we deploy for example, multicloud-gitops, we get the
following errors:

    Cluster level ClusterRoleBinding "vault-server-binding" can not be managed when in namespaced mode⤶
    Cluster level ClusterRoleBinding "golang-external-secrets-controller" can not be managed when in namespaced mode⤶

This happens both when we install the operator via CLI and it ends up in
the pattern-operator-system namespace and when we install it via UI in
the openshift-operators namespace.
Reason for this is that newer Argo versions need and extra variable in
order to be able to set cluster wide permissions.
Let's set ARGOCD_CLUSTER_CONFIG_NAMESPACES like we do via helm

I seem to hit this constantly on 4.9 and a couple of times but not
always(?) on 4.10. We'll stick with '*' for now, this can potentially be
revised later on if needed