Wire the vault unsealing code into the main controller
This commit wires in the three steps to unseal the vault inside the
reconcile loop (initialize, unseal and login). It also adds some rbacs
in order to allow the pattern operator to exec into the vault pod.
Namely we add the following three RBACS:
1) kubebuilder:rbac:groups="",resources=pods,verbs=list;get
   This is needed to check for pods existence
2) kubebuilder:rbac:groups="",resources=pods/exec,resourceNames=vault-0,verbs=create
   This is needed to exec into the vault pod to run commands
3) kubebuilder:rbac:groups="",namespace=patterns-operator-system,resources=secrets,verbs=create;get
   This is needed to create and store the vaultkeys secret inside our
   operator namespace

Note that for 1) and 2) I was unable to restrict them to the vault
namespace. It seems kustomize will barf when the pod resource is being
used twice (with a subcommand /exec) and has a namespace restriction.