Fix subscription update
Currently if a user changes the subscription after the initial pattern
deployment (by tweaking gitOpsSpec.operatorChannel) the subscription
won't be updated. Reason for this are the wrong != nil comparisons
of target.Spec and current.Spec, which will always be true and so
updateSubscription() becomes a noop.

Fix the nil check so that updateSubscription does its job.

Tested on MCG and was able to correctly upgrade gitops from stable to
latest.

Fixes: MBP-351