Switch to gitop-1.8 channel by default
Currently we are using the 'stable' channel which is somewhat
deprecated and is not getting any updates.
Now that we've dropped OCP 4.9 support [1][2] we can switch to the gitops-1.8
channel as the new default gitops version for patterns.
We also hold off releasing the new VP operator version (0.0.12) until
gitops-1.8.2 is released which fixes the login issue with our
name-spaced argo configuration.

Tested on an MCG deployment and correctly got gitops-1.8.1 installed.

[1] https://github.com/hybrid-cloud-patterns/utilities/pull/50
[2] https://github.com/hybrid-cloud-patterns/utilities/pull/51