Get make catalog-build working on ARM-based MacOS
make catalog-build had issues on MacOS because it would try to build an
darwin/arm64 container image when building the OLM catalog. Using the
SQLite-based catalog commands, there is no way to set the platform of
the container image that opm builds.

I converted the catalog-build process to use a file-based catalog
instead of a SQLite-based catalog. That allows us to call docker (or
podman) directly with the --platform tag.

(SQLite-based catalogs are deprecated so this would have to be done
eventually anyways.)