Make sure extraParams are passed down to all the applications
Currently when you pass extraParams like follows:
--set main.extraParameters[0].name=clusterGroup.subscriptions.acm.source
--set main.extraParameters[0].value=iib-${IIB}

These parameters make it to the main argo application installed into the
cluster-wide argocd instance. The problem is that these extraParams do
not make it to the applications that are created inside the namespaced
argocd instance. We fix this by injecting a list of key values pairs
inside a dictionary called "extraParametersNested" inside the "value"
field of the applications that get installed.

A corresponding change in common/ will unroll the key/value pairs under
extraParametersNested and pass them as parameters.

Testes this with:

  make EXTRA_HELM_OPTS="
    --set main.extraParameters[0].name=clusterGroup.subscriptions.acm.source
    --set main.extraParameters[0].value=iib-${IIB}
    --set main.extraParameters[1].name=clusterGroup.subscriptions.acm.channel
    --set main.extraParameters[1].value=${CHANNEL}" install

I correctly observed that the clusterGroup.subscriptions.acm.source
value made it all the way to the common/acm argo application.