Switch to configV1 ClusterVersions to fetch the version of the cluster
The reason we used the equivalent of `oc get OpenShiftControllerManager/cluster -o jsonpath='{.status.version}'`
was that it was a very simple way to get the *current* version of the
cluster. Since certain deployed OCP clusters do not seem to have the
.status.version, we take the more complex route of parsing the status history
of `oc get clusterversion/version -o yaml`. These fields are always
present in all clusters and it is the same approach the UI is
implementing currently [1][2].

Furthermore the clusterVersionStatus docs at [3] specify the following:

	// desired is the version that the cluster is reconciling towards.
	// If the cluster is not yet fully initialized desired will be set
	// with the information available, which may be an image or a tag.
	// +kubebuilder:validation:Required
	// +required
	Desired Release `json:"desired"`

	// history contains a list of the most recent versions applied to the cluster.
	// This value may be empty during cluster startup, and then will be updated
	// when a new update is being applied. The newest update is first in the
	// list and it is ordered by recency. Updates in the history have state
	// Completed if the rollout completed - if an update was failing or halfway
	// applied the state will be Partial. Only a limited amount of update history
	// is preserved.
	// +optional

So what we do in this patch is to check the history and return the first
version in the history which has its state as completed. Should we not
find any (as it might happen during cluster initialization), we fall
back on returning the ClusterVersion.Status.Desired. This is only for
the les slikely case where the cluster still has not version history, in
such cases (and only in such cases), it seems to be okay to just fall
back to that.

Tested both on a ROSA 4.12.x deployed via ClusterBot (which did show the
issue about missing OpenShiftControllerManager/cluster.status.version
and on a stock 4.13 deployed via install-config.

[1] https://github.com/openshift/console/blob/master/frontend/public/module/k8s/cluster-settings.ts#L239
[2] https://github.com/openshift/console/blob/master/frontend/public/module/k8s/cluster-settings.ts#L247
[3] https://pkg.go.dev/github.com/openshift/api/config/v1#ClusterVersionStatus