Introduce ordering in the UI in the gitSpec form
Currently the ordering of the fields displayed in pattern's creation UI
is somewhat random (it depends on the path which is taken from the
JSON's field and it gets sorted alphabetically).
With operator-sdk version 1.30 a new =order field has been introduced [1][2] so
it is now possible to be explicit with the ordering when generating the
bundle via the operator-sdk.

So let's reorder the elements with the most important ones first.
Tested this and I correctly see the following order in the UI:

    Target Repo - Git repo containing the pattern to deploy. Must use https/http
    Target Revision - Branch, tag, or commit to deploy.  Does not support short-sha's. Default: HEAD
    Origin Repo - Upstream git repo containing the pattern to deploy. Used when in-cluster fork to point to the upstream pattern repository
    Origin Revision - Branch, tag or commit in the upstream git repository. Does not support short-sha's. Default to HEAD
    Poll Interval - Interval in seconds to poll for drifts between origin and target repositories. Default: 180 seconds
    Hostname - Optional. FQDN of the git server if automatic parsing from TargetRepo is broken

This now makes a lot more sense and it is a lot clearer.

[1] https://github.com/operator-framework/operator-sdk/commit/0e772903d7c9c9807579a7bdaba3497b1e99cb74
[2] https://sdk.operatorframework.io/docs/building-operators/golang/references/markers/#usage