Make sure we call .actionPerformed as soon as we tweak the .Status of the CRD
In SendPatternInstallationInfo() and SendPatternInstallationUpdate() we
tweak a .Status field in the *api.Pattern struct if everything went
well. We want to make sure that this change is stored in etcd right
away, so there are no chances of it getting lost due to pod restart and
what not.

Let's use the same mechanism that is used everywhere in the reconcile
loop() after tweaking the status and also do it right after the
SendPatternInstallationInfo/Update() call, in order to minimize the chances of
losing this info.

Tested on MCG and verified we only saw one identify event and one update
event.