Drop defaults for clusterGroup chart
Reason for this is that we do not want the CRD to contain any default
and we let the operator pick a default when not set.

This makes it simpler to change version, because we only need to change
it in the operator and we do not need to update the CRD in common and
in all patterns.

Tested by installing an operator with this change and deploying via UI
(i.e. without a predefined default in the CRD) -> correctly deployed MCG