Use pointer for multiSourceConfig Enabled bool
Lester noticed that disabling multiSourceConfig.Enabled in the UI did
not have the intended effect.

Reason for this is that in one of the first reconcile loop iterations
the code that installs the finalizer gets called:

	if !controllerutil.ContainsFinalizer(instance, api.PatternFinalizer) {
		controllerutil.AddFinalizer(instance, api.PatternFinalizer)
		err = r.Client.Update(context.TODO(), instance)
		return r.actionPerformed(instance, "updated finalizer", err)
	}

The issue seems to be that instance.Spec.MultiSourceConfig.Enabled is
correctly set to "false" before the "r.Client.Update()" call, but right
after that call it is set to "true" again (i.e. its default value).

To me this is a bug somewhere in the pkg controller-runtime package.
I did try updating it, but failed miserably because argocd is forcing
older modules. A mere upgrade to controller-runtime v0.13.1 did not
yield any results and I could not find an open issue or a bug related to
this.

Switching the boolean to a pointer does change things, and everything
seems to be working, so let's take that route for the time being.

Tested with both multisource enabled and disabled and both worked
as expected.

Reported-by: Lester Claudio <claudiol@redhat.com>