Instantiate our own cluster-wide Argo CD instance
This feature is disabled by default. It can be enabled by adding the
`initcontainers` string in the `experimentalCapabilities` Spec of the
CRD.

When enabled, it will take over the cluster-wide argo instances and
add an `initContainers` section in the ArgoCD's `repo` part.
These containers will tell Argo to use the internal OCP CA + the
trusted-ca-bundle CAs.