Drop the default containerImage from annotations
The reason for this is somewhat involved: Whne you switch on the
disconnected mode of building the bundle, operator-sdk does a heuristic
of the needed container images and it finds this annotation and resolves
this image to a hash and adds it in the related images:

2024/08/11 11:28:13 visited file or dir: "bundle/manifests/patterns-operator.clusterserviceversion.yaml"
2024/08/11 11:28:13 Pull spec heuristic: quay.io/hybridcloudpatterns/patterns-operator:0.0.1 looks like a pullspec
2024/08/11 11:28:13 Found pullspec for annotation patterns-operator-0.0.1-annotation: quay.io/hybridcloudpatterns/patterns-operator:0.0.1
2024/08/11 11:28:15 Pull spec heuristic: quay.io/hybridcloudpatterns/patterns-operator:0.0.1 looks like a pullspec
2024/08/11 11:28:15 bundle/manifests/patterns-operator.clusterserviceversion.yaml - Replaced pullspec for annotation patterns-operator-0.0.1-annotation: {quay.io hybridcloudpatterns patterns-operator 0.0.1} -> {quay.io hybridcloudpatterns patterns-operator sha256:92f5e9e077a7766240d7d5afdce2032f030b89d3e4983ae099ef0208cea86723}
2024/08/11 11:28:15 Pull spec heuristic: quay.io/hybridcloudpatterns/patterns-operator@sha256:92f5e9e077a7766240d7d5afdce2032f030b89d3e4983ae099ef0208cea86723 looks like a pullspec
2024/08/11 11:28:15 Pull spec heuristic: quay.io/hybridcloudpatterns/patterns-operator@sha256:92f5e9e077a7766240d7d5afdce2032f030b89d3e4983ae099ef0208cea86723 looks like a pullspec
2024/08/11 11:28:15 bundle/manifests/patterns-operator.clusterserviceversion.yaml - Set relateImage patterns-operator-92f5e9e077a7766240d7d5afdce2032f030b89d3e4983ae099ef0208cea86723-annotation (from annotation patterns-operator-92f5e9e077a7766240d7d5afdce2032f030b89d3e4983ae099ef0208cea86723-annotation): quay.io/hybridcloudpatterns/patterns-operator@sha256:92f5e9e077a7766240d7d5afdce2032f030b89d3e4983ae099ef0208cea86723

So you end up with a bundle that also points to quay.io/hybridcloudpatterns/patterns-operator:0.0.1
Per se this is not a huge deal (one extra old image that gets needlessly
pulled), except that in our case this image is super old and was build
with schemaVersion 1 and that precludes it from being uploaded into most
registries and that breaks the `oc mirror` of this operator. So let's
drop it entirely and rebuild the bundle.